api/requirements.txt
flask==3.0.2
werkzeug==3.0.1
flask-cors==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
boto3==1.34.69
redis==5.0.1
SQLAlchemy==2.0.28
gunicorn==21.2.0
